+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [unreleased]
 ### Fixed
+- Base unit-test-case exception expectation optional message parameter
 - Travis build pulling php:5.3
 - Read Me and Change Log fixes for links, WS fixes and typo for other
   documentation files (.md, comment in .sh)